Separates `RegionalizationComponents` into different folders and files to keep project structure
