+++ src/components/common/Navbar/Navbar.tsx
@@ -5,7 +5,7 @@ import CartToggle from 'src/components/cart/CartToggle'
 import SearchInput from 'src/components/common/SearchInput'
 import Icon from 'src/components/ui/Icon'
 import { ButtonIcon, ButtonSignIn } from 'src/components/ui/Button'
-import { RegionalizationButton } from 'src/components/regionalization/RegionalizationComponents'
+import RegionalizationButton from 'src/components/regionalization/RegionalizationButton'
 import Link from 'src/components/ui/Link'
 import Logo from 'src/components/ui/Logo'
 import SlideOver from 'src/components/ui/SlideOver'
