+++ src/components/regionalization/RegionalizationBar/RegionalizationBar.tsx
@@ -1,5 +1,5 @@
 import type { HTMLAttributes } from 'react'
-import { useRegionalization } from 'src/components/regionalization/RegionalizationProvider'
+import { useModal } from 'src/sdk/ui/modal/Provider'
 import Button from 'src/components/ui/Button'
 import Icon from 'src/components/ui/Icon'
 
@@ -13,11 +13,11 @@ export default function RegionalizationBar({
   classes,
   ...otherProps
 }: RegionalizationBarProps) {
-  const { setIsModalOpen } = useRegionalization()
+  const { openModal } = useModal()
 
   return (
     <div data-fs-regionalization-bar className={classes} {...otherProps}>
-      <Button onClick={() => setIsModalOpen(true)}>
+      <Button onClick={() => openModal()}>
         <Icon name="MapPin" width={24} height={24} />
         {content ? (
           <>
