+++ src/components/regionalization/RegionalizationBar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RegionalizationBar'
