+++ src/components/regionalization/RegionalizationButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RegionalizationButton'
