--- src/components/regionalization/RegionalizationComponents/index.tsx
@@ -1,3 +1,0 @@
-export { default } from './RegionalizationBar'
-export { default as RegionalizationButton } from './RegionalizationButton'
-export { default as RegionalizationInput } from './RegionalizationInput'
