--- src/components/regionalization/RegionalizationComponents/regionalization.scss
@@ -1,54 +1,0 @@
-@import "src/styles/scaffold";
-
-.regionalization-input {
-  // Lighthouse
-  margin-top: var(--fs-spacing-3);
-
-  [data-store-label] {
-    font-family: Arial; /* stylelint-disable-line */ // Prevent layout shift
-  }
-
-  [data-store-input] {
-    border: var(--fs-border-width-default) solid var(--fs-border-color-default);
-    border-radius: var(--fs-border-radius-default);
-
-    // Lighthouse
-    height: var(--fs-control-tap-size); /* stylelint-disable-line */
-    margin-left: var(--fs-spacing-3);
-    font-family: Arial; /* stylelint-disable-line */ // prevent layout shift
-  }
-}
-
-[data-fs-regionalization-bar] {
-  width: 100%;
-
-  [data-fs-button] {
-    width: 100%;
-    padding: var(--fs-spacing-0) var(--fs-spacing-3);
-    color: var(--fs-color-text-display);
-    background-color: var(--fs-color-neutral-0);
-    border: 0;
-  }
-
-  span {
-    &:first-of-type {
-      margin-right: auto;
-      font-weight: var(--fs-text-weight-regular);
-    }
-
-    &:not(:first-of-type) {
-      margin-left: auto;
-      text-decoration: underline;
-    }
-  }
-}
-
-[data-fs-regionalization-button] {
-  padding-right: var(--fs-spacing-1);
-  margin-left: calc(-1 * var(--fs-spacing-2));
-  border-right: var(--fs-border-width-default) solid var(--fs-border-color-light);
-
-  [data-fs-button][data-fs-button-variant="tertiary"] {
-    color: var(--fs-color-text-display);
-  }
-}
