+++ src/styles/global/components.scss
@@ -15,8 +15,10 @@
 @import "src/components/product/OutOfStock/out-of-stock.scss";
 
 // Regionalization
+@import "src/components/regionalization/RegionalizationBar/regionalization-bar.scss";
+@import "src/components/regionalization/RegionalizationButton/regionalization-button.scss";
+@import "src/components/regionalization/RegionalizationInput/regionalization-input.scss";
 @import "src/components/regionalization/RegionalizationModal/regionalization-modal.scss";
-@import "src/components/regionalization/RegionalizationComponents/regionalization.scss";
 
 // Search
 @import "src/components/search/Filter/filter.scss";
